==> script/00-create-data.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r}
devtools::install_github("heike/bulletxtrctr") ## Latest commit d6b41e2
#devtools::install_github("nategarton13/bulletcp") ## Latest commit d1e3cea - updates on 12/7/2018 from Nate

library(bulletxtrctr)
library(tidyverse)
library(ggplot2)
```


# HAMBY44

```{r}
hamby44 <- readRDS("../data/hamby44/hamby44_eval.rda")
#hamby44 <- hamby44 %>% filter(!is.na(crosscuts))
```

```{r}
hamby44 <- hamby44 %>% select(c(source, barrel, bullet, land, PARTITION_ID, crosscuts, ccdata_w_resid, left_groove, right_groove))
head(hamby44)

hamby44 <- hamby44 %>% mutate(ccdata = purrr::map(ccdata_w_resid, .f = function(land){
  new_land <- data.frame(x = land$x, value = land$value, y = 0)
  return(new_land)
}))
```

```{r}
hamby44 <- hamby44 %>% mutate(grooves_rollapply = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "rollapply", ccdata = land, return_plot = F)
  #bulletxtrctr::get_grooves_rollapply(x = land$x, value = land$value, return_plot = F)
}))

hamby44 <- hamby44 %>% mutate(grooves_lassobasic = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "lassobasic", ccdata = land, return_plot = F, pred_cutoff = 0.3)
  #bulletxtrctr::get_grooves_lasso(x = land$x, value = land$value, lasso_method = "basic", return_plot = F)
  }))

hamby44 <- hamby44 %>% mutate(grooves_lassofull = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "lassofull", ccdata = land, return_plot = F, pred_cutoff = 0.34)
  #bulletxtrctr::get_grooves_lasso(x = land$x, value = land$value, lasso_method = "full", return_plot = F)
  }))

hamby44 <- hamby44 %>% mutate(grooves_bcp = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  }))

#head(hamby44)
```

```{r}
hamby44_paper <- hamby44
head(hamby44_paper)

saveRDS(hamby44_paper, file = "../data/hamby44/hamby44_paper.rda")
```

# PHOENIX 
```{r}
phoenix <- readRDS("/opt/phoenix/phoenix.rds")

head(phoenix)

phoenix <- phoenix %>% select(-x3p)

phoenix <- phoenix %>% mutate(ccdata = purrr::map(ccdata, .f = function(df){
  df_new <- df %>% group_by(x) %>% dplyr::summarise(value = mean(value, na.rm = T))
  df_new <- as.data.frame(df_new)
  return(df_new)
}))

```


```{r}
phoenix <- phoenix %>% mutate(grooves_rollapply = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "rollapply", ccdata = land, return_plot = F)
  #bulletxtrctr::get_grooves_rollapply(x = land$x, value = land$value, return_plot = F)
}))

phoenix <- phoenix %>% mutate(grooves_lassobasic = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "lassobasic", ccdata = land, return_plot = F, pred_cutoff = 0.3)
  #bulletxtrctr::get_grooves_lasso(x = land$x, value = land$value, lasso_method = "basic", return_plot = F)
  }))

phoenix <- phoenix %>% mutate(grooves_lassofull = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "lassofull", ccdata = land, return_plot = F, pred_cutoff = 0.34)
  #bulletxtrctr::get_grooves_lasso(x = land$x, value = land$value, lasso_method = "full", return_plot = F)
  }))

phoenix <- phoenix %>% mutate(grooves_bcp = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  }))
```



```{r}
phoenix_paper <- phoenix

head(phoenix_paper)

saveRDS(phoenix_paper, file = "../data/phoenix/phoenix_paper.rda")
```


# PHOENIX BCP RE-DO

```{r}
phoenix <- readRDS("../data/phoenix/phoenix_paper.rda")




phoenix_10 <- phoenix[c(1, 4, 5, 7, 8, 9, 10, 12, 13, 14),]
phoenix_10 <- phoenix_10 %>% mutate(grooves_bcp_new = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F, iter = 20000)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  }))

plot_10 <- function(dataset, bullet_num){
  ggplot() + geom_point(data = dataset$ccdata[[bullet_num]], aes(x = x, y = value)) + 
  geom_vline(xintercept = dataset$grooves_bcp_new[[bullet_num]]$groove[1], color = "blue") + 
  geom_vline(xintercept = dataset$grooves_bcp_new[[bullet_num]]$groove[2], color = "blue") +
  geom_vline(xintercept = dataset$grooves_bcp[[bullet_num]]$groove[1], color = "red") + 
  geom_vline(xintercept = dataset$grooves_bcp[[bullet_num]]$groove[2], color = "red") +
  theme_bw()
}


```


# HOUSTON-TEST

```{r}
houston <- readRDS("/opt/houston-test/houston.rds")

head(houston)

houston <- houston %>% select(-x3p)

houston <- houston %>% mutate(ccdata = purrr::map(ccdata, .f = function(df){
  df_new <- df %>% group_by(x) %>% dplyr::summarise(value = mean(value, na.rm = T))
  df_new <- as.data.frame(df_new)
  return(df_new)
}))


head(houston)
```


```{r}
houston <- houston %>% mutate(grooves_rollapply = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "rollapply", ccdata = land, return_plot = F)
  #bulletxtrctr::get_grooves_rollapply(x = land$x, value = land$value, return_plot = F)
}))

houston <- houston %>% mutate(grooves_lassobasic = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "lassobasic", ccdata = land, return_plot = F, pred_cutoff = 0.3)
  #bulletxtrctr::get_grooves_lasso(x = land$x, value = land$value, lasso_method = "basic", return_plot = F)
  }))

houston <- houston %>% mutate(grooves_lassofull = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "lassofull", ccdata = land, return_plot = F, pred_cutoff = 0.34)
  #bulletxtrctr::get_grooves_lasso(x = land$x, value = land$value, lasso_method = "full", return_plot = F)
  }))

houston <- houston %>% mutate(grooves_bcp = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  }))
```


```{r}

houstontest_paper <- houston
head(houstontest_paper)

saveRDS(houstontest_paper, file = "../data/houston-test/houstontest_paper.rda")
```



# HOUSTON-TEST BCP REDO  

```{r}
houston <- readRDS("../data/houston-test/houstontest_paper.rda")




houston_10 <- houston[c(1, 4, 5, 7, 8, 9, 10, 12, 13, 14),]
houston_10 <- houston_10 %>% mutate(grooves_bcp_new = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F, iter = 100000)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  }))

ggplot() + geom_point(data = houston_10$ccdata[[10]], aes(x = x, y = value)) + 
  geom_vline(xintercept = houston_10$grooves_bcp_new[[10]]$groove[1], color = "blue") + 
  geom_vline(xintercept = houston_10$grooves_bcp_new[[10]]$groove[2], color = "blue") +
  geom_vline(xintercept = houston_10$grooves_bcp[[10]]$groove[1], color = "red") + 
  geom_vline(xintercept = houston_10$grooves_bcp[[10]]$groove[2], color = "red") +
  theme_bw()

```




```{r}
houstontest_paper <- houston
head(houstontest_paper)

saveRDS(houstontest_paper, file = "../data/houston-test/houstontest_paper.rda")
```



## TIMING FOR BCP
```{r}
system.time(hamby44[1:10,] %>% mutate(grooves_bcp = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  })))

system.time(phoenix[1:10,] %>% mutate(grooves_bcp = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  })))

system.time(houston[1:10,] %>% mutate(grooves_bcp = purrr::map(ccdata, .f = function(land){
  bulletxtrctr::cc_locate_grooves(method = "bcp", ccdata = land, return_plot = F)
  #bulletcp::get_grooves_bcp(x = land$x, value = land$value, adjust = 10)
  })))



```
